PR update: 1. Added new files user/sensor/stm32/stm32_i2c_task.c and user/sensor/stm32/stm32_i2c_task.h and made necessary updates in CMakeLists.txt 2. Removed unwanted header includes and commented code 3. Moved task creation related logic to the new 2 files mentioned above
