+++ cli_ftd/user/sensor/stm32/stm32_i2c_task.h
@@ -1,0 +1,6 @@
+#ifndef USER_STM32_I2C_SLAVE_H_
+#define USER_STM32_I2C_SLAVE_H_
+
+extern void stm_i2c_comm_taskCreate(void);
+
+#endif
